refactor: move http client commons in util/web
